Move snapshot canonicalization to execution time
DirectoryFileTree was canonicalizing its root directory
at configuration time, which was unnecessarily expensive if
that tree was not used as part of task execution. It was also
inconsistent, since we didn't do the same when snapshotting
individual files or the contents of a tree. As a result we
missed the case when a file was replaced with a symlink.

Both the performance issue and the broken symlink behavior
were fixed by moving the canonicalization into the file snapshotter.